+++ assets/javascripts/kitten/index.js
@@ -115,7 +115,6 @@ export { RangeSlider } from './components/form/range-slider'
 export { RichTextInput } from './components/form/rich-text-input'
 export { Search } from './components/form/search'
 export { SelectWithState } from './components/form/select-with-state'
-export { Slider } from './components/form/slider'
 export { Switch } from './components/form/switch'
 export { TextInputAndSelect } from './components/form/text-input-and-select'
 export { TextInputWithLimit } from './components/form/text-input-with-limit'
@@ -213,11 +212,6 @@ export { SearchInput } from './components/search/search-input'
 
 // Simulators
 export { LoanSimulator } from './components/simulators/loan-simulator'
-
-// Sliders
-export { SliderBar } from './components/sliders/slider-bar'
-export { SliderTooltip } from './components/sliders/slider-tooltip'
-export { SliderWithTooltipAndPower } from './components/sliders/slider-with-tooltip-and-power'
 
 // Status with Bullet
 export { StatusWithBullet } from './components/status-with-bullet'